requirements.txt
# PatentGuard IP — Python Dependencies
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.0
